updated toolbar styling, refactoring based on feedback from devs
