+++ src/components/inputs/SearchLocationsInput/index.js
@@ -1,0 +1,1 @@
+export default from './SearchLocationsInput'
