+++ src/components/toolbars/BaseToolbar/BaseToolbar.js
@@ -19,6 +19,9 @@ const useStyles = makeStyles(theme => ({
     backgroundColor: theme.palette.primary.main,
     boxShadow: '0px 3px 1px -2px rgba(0,0,0,0.2), 0px 2px 2px 0px rgba(0,0,0,0.14), 0px 1px 5px 0px rgba(0,0,0,0.12)',
     overflow: 'auto',
+    position: 'relative',
+    top: -1,
+    minHeight: 71,
   },
 }))
